Keep a new data observation from stalling on a live connection.
Registering an observation decided whether to park it from a connection
status the caller had sampled beforehand, and joined the membership before
applying that decision. Both left windows in which a reconnection processed
concurrently would pass over the observation — it is not parked yet, so it
reports no need for recovery — after which registration parked it. With the
connection already up, no further transition was due to release it, and the
observation waited indefinitely while its status showed it as waiting for a
connection.

`DataObservationManager` now reads the connection status itself, through a
provider supplied by the client, and performs the membership join, the read,
and the decision together under one lock. A status transition can therefore
only run entirely before or entirely after a registration, and the value the
decision uses is the one current at that moment rather than an older sample.
`ConnectionMonitor` publishes its status before notifying, so a reading taken
under the lock is never older than what the manager has already processed.

An observation registered while the connection is known to be unavailable no
longer starts an initial refresh. It is already scheduled for recovery, so
the request would be spent on a channel known to be down and would move the
observation out of the waiting-for-connection status and back again — a flap
visible to the UI that renders connection feedback.

`ObservationRecoveryCoordinator` is renamed to `DataObservationManager`, and
`RecoverableDataObservation` to `ManagedDataObservation`: the class owns the
scope, membership, initialization and shutdown of a client's observations,
not only their recovery. Both types are internal, so no public API changes.
The KDoc of `DataObservation`, `DataObservationImpl` and the manager now
records why each type exists and what the generation bookkeeping protects.
