+++ client/src/main/kotlin/io/spine/chords/client/DesktopClient.kt
@@ -98,9 +98,12 @@ public class DesktopClient(
         io.spine.client.Client.usingChannel(channel).build()
 
     /**
-     * Coordinates recovery of active data observations after connection loss.
+     * Owns the lifecycle of the data observations created by this client:
+     * their background scope, their reaction to connection changes, and their
+     * shutdown.
      */
-    private val recoveryCoordinator = ObservationRecoveryCoordinator()
+    private val observationManager =
+        DataObservationManager({ connectionStatus.value })
 
     /**
      * Runs best-effort subscription cancellation requests.
@@ -118,7 +121,7 @@ public class DesktopClient(
     private val connectionMonitor = ConnectionMonitor(
         { requestConnection -> channel.getState(requestConnection) },
         { source, callback -> channel.notifyWhenStateChanged(source, callback) },
-        recoveryCoordinator::onConnectionStatusChanged
+        observationManager::onConnectionStatusChanged
     )
 
     override val isOpen: Boolean get() = spineClient.isOpen
@@ -127,7 +130,7 @@ public class DesktopClient(
     override val userId: UserId? get() = user()
 
     init {
-        recoveryCoordinator.start()
+        observationManager.start()
         connectionMonitor.start()
 
         getRuntime().addShutdownHook(Thread {
@@ -139,7 +142,7 @@ public class DesktopClient(
         if (!closed.compareAndSet(false, true)) {
             return
         }
-        recoveryCoordinator.close()
+        observationManager.close()
         connectionMonitor.close()
         cancellationScope.cancel()
         closeChannel()
@@ -327,9 +330,13 @@ public class DesktopClient(
      * Creates and registers a recoverable data observation.
      *
      * The observation is returned as soon as it is registered, carrying
-     * [initialValue] and the [DataObservationStatus.Refreshing] status. Its
-     * initial read and subscription run in the recovery coordinator's scope, so
-     * that a slow or unresponsive server does not block the calling thread.
+     * [initialValue]. Its initial read and subscription run in the recovery
+     * coordinator's scope, so that a slow or unresponsive server does not
+     * block the calling thread.
+     *
+     * When the connection is already known to be unavailable, no initial read
+     * is attempted: the observation is returned waiting for a connection, and
+     * the coordinator refreshes it once the connection is restored.
      */
     private fun <T, U> createObservation(
         initialValue: T,
@@ -346,10 +353,10 @@ public class DesktopClient(
             subscribe,
             applyUpdate,
             { connectionStatus.value },
-            recoveryCoordinator::unregister,
-            recoveryCoordinator::retryWhileConnected
+            observationManager::unregister,
+            observationManager::retryWhileConnected
         )
-        recoveryCoordinator.registerAndInitialize(observation, connectionStatus.value)
+        observationManager.registerAndInitialize(observation)
         return observation
     }
