+++ client/src/test/kotlin/io/spine/chords/client/DataObservationImplSpec.kt
@@ -37,8 +37,15 @@ import kotlinx.coroutines.CancellationException
 import kotlinx.coroutines.Dispatchers.IO
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.runBlocking
+import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
 
+/**
+ * Tests how [DataObservationImpl] publishes data, tracks its status, and
+ * handles cancellation, connection loss, and subscriptions that arrive after
+ * the refresh that created them is no longer current.
+ */
+@DisplayName("`DataObservationImpl` should")
 internal class DataObservationImplSpec {
 
     @Test
@@ -220,6 +227,10 @@ internal class DataObservationImplSpec {
         observation.status.value shouldBe DataObservationStatus.Active
     }
 
+    /**
+     * An initial refresh that cannot reach the server leaves the observation
+     * waiting rather than failed, so a later refresh still brings in the data.
+     */
     @Test
     fun `recover after the initial refresh cannot connect`() {
         val source = FakeObservationSource("server data")
@@ -244,6 +255,11 @@ internal class DataObservationImplSpec {
         observation.status.value shouldBe DataObservationStatus.Active
     }
 
+    /**
+     * An observation parked before it ever refreshed — the state in which a
+     * disconnected client hands one out — still reads and subscribes normally
+     * once the connection is restored.
+     */
     @Test
     fun `recover after waiting for connection before the first refresh`() {
         val source = FakeObservationSource("server data")
@@ -415,6 +431,11 @@ internal class DataObservationImplSpec {
         cancelCalls.get() shouldBe 0
     }
 
+    /**
+     * The initial value is readable while the first refresh is still creating
+     * its subscription. This is what lets observation creation return without
+     * waiting for the server.
+     */
     @Test
     fun `expose the initial value while the first refresh is pending`(): Unit = runBlocking {
         val subscribeStarted = CountDownLatch(1)
@@ -442,6 +463,11 @@ internal class DataObservationImplSpec {
         observation.status.value shouldBe DataObservationStatus.Active
     }
 
+    /**
+     * A subscription arriving after the client closed the observation is left
+     * alone: the whole channel is going away, so cancelling the subscription
+     * individually would be a pointless call on a dying connection.
+     */
     @Test
     fun `not cancel subscription created after the client is closed`(): Unit = runBlocking {
         val subscribeStarted = CountDownLatch(1)
@@ -543,8 +569,8 @@ private class FakeObservationSource(
     fun createObservation(
         initialValue: String = "",
         connectionStatus: () -> ConnectionStatus = { ConnectionStatus.CONNECTED },
-        onRecoveryNeeded: (RecoverableDataObservation) -> Unit = {},
-        onCancelled: (RecoverableDataObservation) -> Unit = {}
+        onRecoveryNeeded: (ManagedDataObservation) -> Unit = {},
+        onCancelled: (ManagedDataObservation) -> Unit = {}
     ): DataObservationImpl<String, String> = DataObservationImpl(
         initialValue,
         {
